// src/constants/telegram/enums/state-status.enum.ts
export enum StateStatuses {
  WAITING_FOR_DOCUMENTS = 'WAITING_FOR_DOCUMENTS',
  DOCUMENTS_RECEIVED = 'DOCUMENTS_RECEIVED',
  WAITING_FOR_PRICE_CONFIRMATION = 'WAITING_FOR_PRICE_CONFIRMATION',
  COMPLETED = 'COMPLETED',
  FINAL_STEP = 'FINAL_STEP',
  WAITING_FOR_DATA_CONFIRMATION = 'WAITING_FOR_DATA_CONFIRMATION',
}