// src/constants/telegram/enums/bot-answers.enum.ts
export enum BOT_MESSAGES {
  PRICE_CONFIRMATION = 'Вартість страхування становить 100 USD. Ви погоджуєтесь?',
  REUPLOAD_DOCUMENTS = 'Будь ласка, повторно завантажте документи, якщо дані виявились некоректними.',
}